requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.1.6
idna==3.4
paypalrestsdk==1.13.1
Pillow==10.0.0
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.2.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.0.4